refactor(bot): modularize slash commands into cogs and enhance logging colors
Transformed all slash commands into separate cogs for better organization. Introduced a dynamic command handler in the main file. Also updated the console logging colors for improved visibility.
